move lock to prevent circular imports
